+++ flask_xxljob/__init__.py
@@ -34,8 +34,7 @@ execute the actual tasks.
 
 from __future__ import annotations
 
-__version__ = "0.2.0"
-
+from ._version import __version__
 from .client import AdminCallResult, CallResult
 from .exceptions import (
     FlaskXXLJobError,